requirements.txt
numpy
pandas
requests
torch
torchaudio
transformers
sentence-transformers
pyannote.audio
openai-whisper
scikit-learn
plotly
pydub
wave
librosa
soundfile
ffmpeg-python
datetime
sounddevice
matplotlib
